+++ scripts/v1/train/save.py
@@ -83,9 +83,12 @@ def save_gpqr(
     likelihood,
     model,
     path,
+    *,
+    quantile_gap_lower_bound,
 ):
     data = dict(
         quantiles=quantiles,
+        quantile_gap_lower_bound=float(quantile_gap_lower_bound),
         X_scaler=dict(
             type=X_scaler.__class__.__name__,
             args=dict(dim=X_scaler.dim, batch_shape=X_scaler.batch_shape),
@@ -99,9 +102,8 @@ def save_gpqr(
         likelihood=dict(
             type=likelihood.__class__.__name__,
             args=dict(
-                quantile_levels=likelihood.quantile_levels,
-                central_quantile_idx=likelihood.central_quantile_idx,
-                quantile_slope_lower_bound=likelihood.quantile_slope_lower_bound,
+                quantile_levels=quantiles,
+                central_quantile_idx=model.num_lower_quantiles[0],
                 **_prior_args(likelihood, "noise"),
                 batch_shape=likelihood.batch_shape,
             ),
@@ -114,8 +116,7 @@ def save_gpqr(
                 num_quantiles=model.num_quantiles[0],
                 num_lower_quantiles=model.num_lower_quantiles[0],
                 num_latents=model.num_latents,
-                quantile_levels=model.quantile_levels,
-                quantile_slope_lower_bound=model.quantile_slope_lower_bound,
+                quantile_levels=quantiles,
                 **_prior_args(model, "lengthscale"),
                 fixed_lengthscale=(
                     model.covar_module.base_kernel.lengthscale.detach().clone()